feat(auth): move auth routes to dedicated class
- Created AuthRoutes class in features/common/auth/auth.routes.ts
- Moved auth routes from index.ts to dedicated class
- Registered AuthRoutes in DI container
- Updated index.ts to use AuthRoutes via container resolution
- Preserved all existing routes functionality
